+++ src/main/java/com/denise/portfolio/models/ExperienciaModel.java
@@ -3,7 +3,7 @@ package com.denise.portfolio.models;
 import com.fasterxml.jackson.annotation.JsonFormat;
 import com.fasterxml.jackson.annotation.JsonFormat.Shape;
 
-import jakarta.persistence.*;
+import javax.persistence.*;
 
 @Entity
 @Table(name="experiencia")
@@ -12,7 +12,7 @@ public class ExperienciaModel {
 	@Id
 	@GeneratedValue(strategy=GenerationType.IDENTITY)
 	@Column(unique=true, nullable=false)
-	private Integer id;
+	private Long id;
 	private String titulo;	
 	private String logo;
 	private String descripcion;
@@ -22,10 +22,10 @@ public class ExperienciaModel {
 	@JsonFormat(pattern = "yyyy-MM-dd", shape = Shape.STRING)
 	@Column(name="fin")
 	private String fin;
-	public Integer getId() {
+	public Long getId() {
 		return id;
 	}
-	public void setId(Integer id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
 	public String getTitulo() {
