+++ src/main/java/com/denise/portfolio/models/PersonaModel.java
@@ -3,7 +3,7 @@ package com.denise.portfolio.models;
 import com.fasterxml.jackson.annotation.JsonFormat;
 import com.fasterxml.jackson.annotation.JsonFormat.Shape;
 
-import jakarta.persistence.*;
+import javax.persistence.*;
 
 @Entity
 @Table(name="persona")
@@ -11,7 +11,7 @@ public class PersonaModel {
 	@Id
 	@GeneratedValue(strategy=GenerationType.IDENTITY)
 	@Column(unique=true, nullable=false)
-	private Integer id;
+	private Long id;
 	private String nombre;
 	private String apellido;
 	@JsonFormat(pattern="yyyy-MM-dd", shape=Shape.STRING)
@@ -24,10 +24,10 @@ public class PersonaModel {
 	private String imagen_fondo;
 	private String imagen_perfil;
 	
-	public Integer getId() {
+	public Long getId() {
 		return id;
 	}
-	public void setId(Integer id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
 	public String getNombre() {
