+++ src/main/java/com/denise/portfolio/models/ProyectoModel.java
@@ -2,7 +2,7 @@ package com.denise.portfolio.models;
 import com.fasterxml.jackson.annotation.JsonFormat;
 import com.fasterxml.jackson.annotation.JsonFormat.Shape;
 
-import jakarta.persistence.*;
+import javax.persistence.*;
 
 @Entity
 @Table(name="proyectos")
@@ -10,11 +10,11 @@ public class ProyectoModel {
 	@Id
 	@GeneratedValue(strategy=GenerationType.IDENTITY)
 	@Column(unique = true, nullable = false)
-	private Integer id;
-	public Integer getId() {
+	private Long id;
+	public Long getId() {
 		return id;
 	}
-	public void setId(Integer id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
 	public String getTitulo() {
